+++ include/NbfmDecode.h
@@ -50,20 +50,20 @@ public:
   void process(const IQSampleVector &samples_in, SampleVector &audio);
 
   /** Return actual frequency offset in Hz with respect to receiver LO. */
-  double get_tuning_offset() const { return m_baseband_mean * freq_dev; }
+  float get_tuning_offset() const { return m_baseband_mean * freq_dev; }
 
   /** Return RMS baseband signal level (where nominal level is 0.707). */
-  double get_baseband_level() const { return m_baseband_level; }
+  float get_baseband_level() const { return m_baseband_level; }
 
   // Return RMS IF level.
-  double get_if_rms() const { return m_if_rms; }
+  float get_if_rms() const { return m_if_rms; }
 
 private:
   // Data members.
   const double m_sample_rate_fmdemod;
-  double m_baseband_mean;
-  double m_baseband_level;
-  double m_if_rms;
+  float m_baseband_mean;
+  float m_baseband_level;
+  float m_if_rms;
 
   SampleVector m_buf_baseband;
   SampleVector m_buf_baseband_raw;
