+++ sfmbase/FmDecode.cpp
@@ -314,7 +314,7 @@ void FmDecoder::process(const IQSampleVector &samples_in, SampleVector &audio) {
   }
 
   // Measure baseband level.
-  double baseband_mean, baseband_rms;
+  float baseband_mean, baseband_rms;
   Utility::samples_mean_rms(m_buf_baseband, baseband_mean, baseband_rms);
   m_baseband_mean = 0.95 * m_baseband_mean + 0.05 * baseband_mean;
   m_baseband_level = 0.95 * m_baseband_level + 0.05 * baseband_rms;
